Feat: 管理ページ作成

+++ backend/app/controllers/api/v1/users_controller.rb
@@ -1,30 +1,45 @@
 module Api
   module V1
     class UsersController < ApplicationController
-      before_action :authenticate_api_v1_user!
       def index
-        users = User.all
-        render json: users
+        if current_api_v1_user.admin?
+          users = User.all
+          render json: users
+        else
+          render json: { message: '権限がありません', success: false }, status: :unauthorized
+        end
       end
 
       def show
-        user = User.find(params[:id])
-        render json: user
+        if current_or_admin_user
+          user = User.find(params[:id])
+          render json: user
+        else
+          render json: { message: '権限がありません', success: false }, status: :unauthorized
+        end
       end
 
       def update
-        user = User.find(params[:id])
-        if user.update(user_params)
-          render json: user
+        if current_or_admin_user
+          user = User.find(params[:id])
+          if user.update(user_params)
+            render json: user
+          else
+            render json: { errors: user.errors }, status: :unprocessable_entity
+          end
         else
-          render json: { errors: user.errors }, status: :unprocessable_entity
+          render json: { message: '権限がありません', success: false }, status: :unauthorized
         end
       end
 
       def destroy
-        user = User.find(params[:id])
-        user.destroy
-        render json: { message: 'ユーザーが削除されました', success: true }
+        if current_or_admin_user
+          user = User.find(params[:id])
+          user.destroy
+          render json: { message: 'ユーザーが削除されました', success: true }
+        else
+          render json: { message: '権限がありません', success: false }, status: :unauthorized
+        end
       end
 
       private
